+++ src/Request.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Mediawiki\Api;
+
+/**
+ * @since 0.2
+ */
+interface Request {
+
+	/**
+	 * @since 0.2
+	 * @return array
+	 */
+	public function getParams();
+
+}