+++ RELEASENOTES.md
@@ -3,6 +3,9 @@ These are the release notes for the [mediawiki-api-base](README.md).
 ## Version 0.2 (under development)
 
 * Unsuccessful logins now throw a UsageException with extra details
+* The Request interface and SimpleRequest class have been added
+* MediawikiApi now have a getRequest and postRequest object
+* MediawikiApi getAction and postAction methods have been deprecated in favour of the above
 
 ## Version 0.1.2 (25 May 2014)